src/requirements.txt
altair==4.2.0
numpy==1.22.3
pandas==1.4.2
Pillow==9.3.0
python_dateutil==2.8.2
streamlit==1.14.0
yfinance==0.1.70
